refactor: move data-manager
