+++ app/javascript/dashboard/helper/CacheHelper/data-manager.js
