+++ app/javascript/dashboard/helper/CacheHelper/version.js
